data processing methods in separate file
